refactor: use metadata API
