+++ migration/portfolio_v2.5/src/app/layout.tsx
@@ -3,22 +3,13 @@
 import "./globals.css";
 import "./_old-styles_/_Pages.css";
 import "./_old-styles_/Homepage.css";
-import type { Metadata } from "next";
-import { Geist, Geist_Mono } from "next/font/google";
 import { useContext, useEffect } from "react";
 import Header from "../Components/Header/Header";
 import Footer from "../Components/Footer/Footer";
 import ClientProvider from "../Components/ClientProvider";
 import { ThemeContext } from "../context/ThemeContext";
 
-const geistSans = Geist({ variable: "--font-geist-sans", subsets: ["latin"] });
-const geistMono = Geist_Mono({ variable: "--font-geist-mono", subsets: ["latin"] });
 
-const metadata: Metadata = {
-  title: "Kainen White | Product & UX Designer",
-  description:
-    "The design portfolio of Kainen White, a product and UX designer creating user-centered digital experiences that drive business growth. View case studies.",
-};
 
 export default function RootLayout({ children }: { children: React.ReactNode }) {
   const { theme } = useContext(ThemeContext);
@@ -30,7 +21,7 @@ export default function RootLayout({ children }: { children: React.ReactNode }) 
 
   return (
     <html lang="en" data-theme={theme}>
-      <body className={`${geistSans.variable} ${geistMono.variable} antialiased`}>
+      <body className="antialiased">
         <ClientProvider>
           <a href="#main-content" className="skip-link">
             Skip to main content
